+++ IdentityServer/EC.IdentityServer/Quickstart/Device/DeviceController.cs
@@ -6,6 +6,9 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using EC.IdentityServer.Quickstart.Models;
+using EC.IdentityServer.Quickstart.Options;
+using EC.IdentityServer.Quickstart.ViewModels;
 using IdentityServer4.Configuration;
 using IdentityServer4.Events;
 using IdentityServer4.Extensions;
