+++ IdentityServer/EC.IdentityServer/Quickstart/Home/HomeController.cs
@@ -2,6 +2,7 @@
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
 
+using EC.IdentityServer.Quickstart.ViewModels;
 using IdentityServer4.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Hosting;
