+++ IdentityServer/EC.IdentityServer/Quickstart/Models/ConsentInputModel.cs
@@ -4,7 +4,7 @@
 
 using System.Collections.Generic;
 
-namespace IdentityServerHost.Quickstart.UI
+namespace EC.IdentityServer.Quickstart.Models
 {
     public class ConsentInputModel
     {
