+++ IdentityServer/EC.IdentityServer/Quickstart/Models/LogoutInputModel.cs
@@ -1,12 +1,10 @@
 ﻿// Copyright (c) Brock Allen & Dominick Baier. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
-
-namespace IdentityServerHost.Quickstart.UI
+namespace EC.IdentityServer.Quickstart.Models
 {
-    public class ExternalProvider
+    public class LogoutInputModel
     {
-        public string DisplayName { get; set; }
-        public string AuthenticationScheme { get; set; }
+        public string LogoutId { get; set; }
     }
 }
