+++ IdentityServer/EC.IdentityServer/Quickstart/Options/ConsentOptions.cs
@@ -1,8 +1,7 @@
 ﻿// Copyright (c) Brock Allen & Dominick Baier. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
-
-namespace IdentityServerHost.Quickstart.UI
+namespace EC.IdentityServer.Quickstart.Options
 {
     public class ConsentOptions
     {
