+++ IdentityServer/EC.IdentityServer/Quickstart/Providers/ExternalProvider.cs
@@ -1,11 +1,11 @@
 ﻿// Copyright (c) Brock Allen & Dominick Baier. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
-
-namespace IdentityServerHost.Quickstart.UI
+namespace EC.IdentityServer.Quickstart.Providers
 {
-    public class LogoutViewModel : LogoutInputModel
+    public class ExternalProvider
     {
-        public bool ShowLogoutPrompt { get; set; } = true;
+        public string DisplayName { get; set; }
+        public string AuthenticationScheme { get; set; }
     }
 }
