+++ IdentityServer/EC.IdentityServer/Quickstart/ViewModels/DiagnosticsViewModel.cs
@@ -8,7 +8,7 @@ using Newtonsoft.Json;
 using System.Collections.Generic;
 using System.Text;
 
-namespace IdentityServerHost.Quickstart.UI
+namespace EC.IdentityServer.Quickstart.ViewModels
 {
     public class DiagnosticsViewModel
     {
