+++ IdentityServer/EC.IdentityServer/Quickstart/ViewModels/LogoutViewModel.cs
@@ -1,0 +1,13 @@
+﻿// Copyright (c) Brock Allen & Dominick Baier. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
+
+
+using EC.IdentityServer.Quickstart.Models;
+
+namespace EC.IdentityServer.Quickstart.ViewModels
+{
+    public class LogoutViewModel : LogoutInputModel
+    {
+        public bool ShowLogoutPrompt { get; set; } = true;
+    }
+}
